feat: Deadline-based escalation system, PDF overlay fixes, fax/email improvements
- Escalation: deadline reached → action_needed + manager notification
- Escalation: deadline+14d → admin notification + email
- Remove follow-up count based escalation, use deadline-based only
- Add action_needed to case_providers overall_status ENUM
- PDF overlay: fix auto page break issue, add coordinate fine-tune inputs
- Fax helper: add Faxage integration support
- PDF generator: DomPDF HTML→PDF conversion
- Case detail: split into partial components
- Health tracker: replace health ledger with new tracker
- MR tracker: new medical records tracker page
- Frontend: refactored providers, cases, admin pages
- Shared JS utilities: listPageBase, scrollContainer, sortIcon

+++ backend/api/documents/upload.php
@@ -35,18 +35,18 @@ $notes = $_POST['notes'] ?? null;
 
 // Provider template fields
 $isProviderTemplate = isset($_POST['is_provider_template']) && $_POST['is_provider_template'] == '1';
-$providerNameX = !empty($_POST['provider_name_x']) ? (int)$_POST['provider_name_x'] : null;
-$providerNameY = !empty($_POST['provider_name_y']) ? (int)$_POST['provider_name_y'] : null;
-$providerNameWidth = !empty($_POST['provider_name_width']) ? (int)$_POST['provider_name_width'] : null;
-$providerNameHeight = !empty($_POST['provider_name_height']) ? (int)$_POST['provider_name_height'] : null;
+$providerNameX = !empty($_POST['provider_name_x']) ? (float)$_POST['provider_name_x'] : null;
+$providerNameY = !empty($_POST['provider_name_y']) ? (float)$_POST['provider_name_y'] : null;
+$providerNameWidth = !empty($_POST['provider_name_width']) ? (float)$_POST['provider_name_width'] : null;
+$providerNameHeight = !empty($_POST['provider_name_height']) ? (float)$_POST['provider_name_height'] : null;
 $providerNameFontSize = !empty($_POST['provider_name_font_size']) ? (int)$_POST['provider_name_font_size'] : 12;
 
 // Date overlay fields
 $useDateOverlay = isset($_POST['use_date_overlay']) && $_POST['use_date_overlay'] == '1';
-$dateX = !empty($_POST['date_x']) ? (int)$_POST['date_x'] : null;
-$dateY = !empty($_POST['date_y']) ? (int)$_POST['date_y'] : null;
-$dateWidth = !empty($_POST['date_width']) ? (int)$_POST['date_width'] : null;
-$dateHeight = !empty($_POST['date_height']) ? (int)$_POST['date_height'] : null;
+$dateX = !empty($_POST['date_x']) ? (float)$_POST['date_x'] : null;
+$dateY = !empty($_POST['date_y']) ? (float)$_POST['date_y'] : null;
+$dateWidth = !empty($_POST['date_width']) ? (float)$_POST['date_width'] : null;
+$dateHeight = !empty($_POST['date_height']) ? (float)$_POST['date_height'] : null;
 $dateFontSize = !empty($_POST['date_font_size']) ? (int)$_POST['date_font_size'] : 12;
 
 // Validate document type
